Lib cleanup: singular action.ts, move search action, flatten single-file dirs
Three small convention fixes in lib/, no behavior changes:

1. Rename lib/collections/actions.ts → lib/collections/action.ts. Singular
   matches lib/cart/action.ts and lib/search/action.ts.
2. Move loadMoreSearchProducts from lib/collections/action.ts to
   lib/search/action.ts where it conceptually belongs. Update the two
   importers in components/{search,collections}.
3. Collapse lib/utils/index.ts → lib/utils.ts and lib/product/index.ts →
   lib/product.ts. Each directory contained only index.ts; the rest of
   lib/ is flat for files of this kind. Importers unchanged — @/lib/utils
   and @/lib/product resolve to the flat file the same way.

Adds a template-rollout-log entry so downstream storefronts can adopt.
