+++ apps/template/components/collections/results-grid.tsx
@@ -2,7 +2,7 @@ import { getTranslations } from "next-intl/server";
 import { Suspense } from "react";
 
 import { ProductCard, ProductCardSkeleton } from "@/components/product-card/product-card";
-import { loadMoreCollectionProducts } from "@/lib/collections/actions";
+import { loadMoreCollectionProducts } from "@/lib/collections/action";
 import type { CollectionResultsData } from "@/lib/collections/server";
 import type { Locale } from "@/lib/i18n";
