+++ apps/template/lib/product.ts
