+++ apps/template/lib/search/action.ts
@@ -1,7 +1,9 @@
 "use server";
 
+import { getProducts } from "@/lib/shopify/operations/products";
 import { predictiveSearch } from "@/lib/shopify/operations/search";
-import type { PredictiveSearchResult } from "@/lib/types";
+import type { ProductFilter } from "@/lib/shopify/types/filters";
+import type { PageInfo, PredictiveSearchResult, ProductCard } from "@/lib/types";
 
 export async function predictiveSearchAction(
   query: string,
@@ -12,4 +14,28 @@ export async function predictiveSearchAction(
   }
 
   return predictiveSearch(query.trim(), locale, 4);
+}
+
+export async function loadMoreSearchProducts(params: {
+  query?: string;
+  collection?: string;
+  cursor: string;
+  sortKey?: string;
+  filters?: ProductFilter[];
+  locale: string;
+}): Promise<{ products: ProductCard[]; pageInfo: PageInfo; total: number }> {
+  const result = await getProducts({
+    query: params.query,
+    collection: params.collection,
+    cursor: params.cursor,
+    sortKey: params.sortKey,
+    filters: params.filters,
+    locale: params.locale,
+  });
+
+  return {
+    products: result.products,
+    pageInfo: result.pageInfo,
+    total: result.total,
+  };
 }
