+++ apps/template/lib/utils.ts
